refactor: move pipe_roughness to friction.h/cpp
pipe_roughness and standard_pipe_roughness belong alongside the friction
factor correlations - users call pipe_roughness() to get e_D for
friction_haaland/serghides/colebrook. References are identical (Moody,
Colebrook, White).

- friction.h: add pipe_roughness + standard_pipe_roughness declarations
- friction.cpp: move implementation from utils.cpp
- utils.h: remove roughness declarations + unused includes
- utils.cpp: remove roughness implementation + unused includes
